OPENEUROPA-1494: Minor refactor of the Syncope-down graceful handling.
