+++ modules/oe_authorisation_syncope/oe_authorisation_syncope.module
@@ -11,6 +11,7 @@ use Drupal\Core\Entity\EntityInterface;
 use Drupal\Core\Entity\EntityTypeInterface;
 use Drupal\Core\Field\BaseFieldDefinition;
 use Drupal\Core\Form\FormStateInterface;
+use Drupal\oe_authorisation_syncope\Exception\SyncopeDownException;
 use Drupal\oe_authorisation_syncope\Exception\SyncopeUserException;
 use Drupal\user\RoleInterface;
 use Drupal\user\UserInterface;
@@ -19,11 +20,6 @@ use Drupal\user\UserInterface;
  * Implements hook_entity_presave().
  */
 function oe_authorisation_syncope_entity_presave(EntityInterface $entity) {
-
-  if (oe_authorisation_syncope_disabled()) {
-    return;
-  }
-
   if ($entity instanceof RoleInterface) {
     \Drupal::service('oe_authorisation_syncope.role_mapper')
       ->preSave($entity);
@@ -41,11 +37,6 @@ function oe_authorisation_syncope_entity_presave(EntityInterface $entity) {
  * Implements hook_entity_predelete().
  */
 function oe_authorisation_syncope_entity_predelete(EntityInterface $entity) {
-
-  if (oe_authorisation_syncope_disabled()) {
-    return;
-  }
-
   if ($entity instanceof RoleInterface) {
     \Drupal::service('oe_authorisation_syncope.role_mapper')->preDelete($entity);
     return;
@@ -68,11 +59,6 @@ function oe_authorisation_syncope_user_login($account) {
  * Implements hook_ENTITY_TYPE_storage_load() for the User entity type.
  */
 function oe_authorisation_syncope_user_storage_load(array $entities) {
-
-  if (oe_authorisation_syncope_disabled()) {
-    return;
-  }
-
   foreach ($entities as $entity) {
     \Drupal::service('oe_authorisation_syncope.user_mapper')->load($entity);
   }
@@ -106,8 +92,7 @@ function oe_authorisation_syncope_entity_base_field_info(EntityTypeInterface $en
  * an up-to-date status of the roles as they are in Syncope.
  */
 function oe_authorisation_syncope_form_user_form_alter(&$form, FormStateInterface $form_state): void {
-
-  if (oe_authorisation_syncope_disabled()) {
+  if (!\Drupal::service('oe_authorisation_syncope.client')::isEnabled()) {
     return;
   }
 
@@ -135,15 +120,8 @@ function oe_authorisation_syncope_form_user_form_alter(&$form, FormStateInterfac
     \Drupal::messenger()->addError('There is a problem with the authorisation service. User roles cannot be updated.');
     $form['account']['roles']['#disabled'] = TRUE;
   }
-}
-
-/**
- * Checks if Syncope integration is disabled by config override.
- */
-function oe_authorisation_syncope_disabled() {
-  $oe_authorisation_syncope_settings = \Drupal::config('oe_authorisation_syncope.settings');
-
-  if (!empty($oe_authorisation_syncope_settings->get('disabled'))) {
-    return $oe_authorisation_syncope_settings->get('disabled');
+  catch (SyncopeDownException $e) {
+    \Drupal::messenger()->addError($e->getMessage());
+    $form['account']['roles']['#disabled'] = TRUE;
   }
 }
