+++ modules/oe_authorisation_syncope/oe_authorisation_syncope.services.yml
@@ -9,7 +9,7 @@ services:
     class: Drupal\oe_authorisation_syncope\SyncopeUserMapper
     arguments: ['@oe_authorisation_syncope.client', '@oe_authorisation_syncope.role_mapper', '@logger.factory', '@entity_type.manager']
   oe_authorisation_syncope.subscriber:
-    class: Drupal\oe_authorisation_syncope\EventSubscriber\OeAuthorisationSyncopeSubscriber
+    class: Drupal\oe_authorisation_syncope\EventSubscriber\SyncopeExceptionSubscriber
     arguments:
       - '@messenger'
     tags:
